requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
Pillow>=8.0.0
moviepy>=1.0.3
pyttsx3>=2.90
gTTS>=2.2.0 